Title(feat): 完成播放器控件,底部进度条效果。 Cause：N/A Solution：N/A Effect: N/A。 Reviewer: N/A 附加说明：N/A

+++ module-Music/src/main/java/com/zeke/music/fragments/VodInfoFragment.kt
@@ -225,7 +225,7 @@ class VodInfoFragment : BaseFragment(), IPlayerView {
     override fun showPauseStateView() {}
     override fun showFlowTipsView() {}
     override fun switchVisibleState() {}
-    override fun updatePlayProgressView(isDrag: Boolean, postion: Int) {}
+    override fun updatePlayProgressView(isDrag: Boolean, postion: Long) {}
     override fun repostControllersDismissTask(enable: Boolean) {}
     override fun dismissPop() {}
 }
